requirements.txt
timm
flash_attn
einops
transformers
pillow
requests
flask
ultralytics